Some Packets
Mostly outgoing packets are now able to be written to stream.

+++ src/icac/impervium/server/networking/Packet.java
@@ -1,5 +1,7 @@
 package icac.impervium.server.networking;
 
+import icac.impervium.server.datatypes.UInt8;
+
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 
@@ -9,7 +11,7 @@ public abstract class Packet {
 		
 	}
 	
-	public abstract Integer getID();
+	public abstract UInt8 getID();
 	public abstract void write(DataOutputStream dos) throws Exception;
 	public abstract void read(DataInputStream dis) throws Exception;
 }
